+++ tests/plugin_test/src/191_hashes.php
@@ -1,0 +1,8 @@
+<?php
+
+// Suppress warning in php 7 since php-ast doesn't provide this
+#This is a comment
+function test191() {
+    echo "In test\n";
+}
+test191();#